Fix incompatibility with v5 branch (#15)
In the v5 branch of the SDK, we added strict type hints throughout the
code base. This uncovered an error in our redis implementation where we
were re-defining the `RedisFeatureRequester::_options` array.

The solution is a simple property rename to avoid the conflict.

In order to support both the 4.x and 5.x branches of the SDK, we had to
loosen the type signature on our `makeRequester` method. We also made
some modifications to the way the shared tests are pulled into the CI process.